py-service-monitor-sefaz/requirements.txt
requests
beautifulsoup4
python-dotenv
Flask
